requirements.txt
# --- Streamlit Core ---
streamlit>=1.35.0
streamlit-tags>=1.2.1

# --- Web Scraping & Automation ---
selenium>=4.21.0
webdriver-manager>=4.0.1
requests>=2.31.0
html2text>=2020.1.16
beautifulsoup4>=4.12.3
playwright>=1.44.0

# --- Data Handling ---
pandas>=2.2.2
XlsxWriter>=3.2.0
tabulate>=0.9.0             # ✅ Required by pandas for pretty display and output

# --- Google Generative AI ---
google-generativeai>=0.3.2

# --- Optional: OpenAI (Only if needed) ---
openai>=1.30.1              # ⚠️ Only keep if you're planning to use OpenAI

# --- Environment & Validation ---
python-dotenv>=1.0.1
pydantic>=2.7.1

# --- Typing support ---
typing-extensions>=4.10.0
